requirements.txt
numpy                           ==1.22.4
django                          ==3.1.12
djangorestframework             ==3.12.4
# django-cors-headers             =3.2.1
requests                        ==2.26.0
PyYAML                          ==6.0
pymongo[snappy,gssapi,srv,tls]  ==3.10.1
# dnspython                       ==2.1.0
djongo                          ==1.3.6
django-crontab                  ==0.7.1
django-extensions               ==3.1.5
